10 src/1010main/epic3_certification_roadmap/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.3.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0 